// src/api/frame/tags/EPTag.ts
export enum EPTag {
	REQ_IS_READY_FOR_SWITCH = '0B000003',
	REQ_IS_GRID_CONNECTED = '0B000004',
	REQ_IS_ISLAND_GRID = '0B000005',
	REQ_IS_INVALID_STATE = '0B000006',
	REQ_IS_POSSIBLE = '0B000007',
	IS_READY_FOR_SWITCH = '0B800003',
	IS_GRID_CONNECTED = '0B800004',
	IS_ISLAND_GRID = '0B800005',
	IS_INVALID_STATE = '0B800006',
	IS_POSSIBLE = '0B800007',
	GENERAL_ERROR = '0BFFFFFF',
	REQ_SWITCH_TO_GRID = '0B000001',
	REQ_SWITCH_TO_ISLAND = '0B000002',
	REQ_LEAVE_INVALID_STATE_TO_ISLAND = '0B000008',
	REQ_LEAVE_INVALID_STATE_TO_GRID = '0B000009',
	SWITCH_TO_GRID = '0B800001',
	SWITCH_TO_ISLAND = '0B000009',
	LEAVE_INVALID_STATE_TO_ISLAND = '0B800008',
	LEAVE_INVALID_STATE_TO_GRID = '0B800009',
	REQ_EP_RESERVE = '1B000009',
	PARAM_INDEX = '1B040000',
	PARAM_INT = '1B040011',
	PARAM_EP_RESERVE = '1B040023',
	PARAM_TIME_LAST_FULL = '1B040024',
	PARAM_TIME_LAST_EMPTY = '1B040025',
	PARAM_LAST_SOC = '1B040026',
	PARAM_EP_RESERVE_ENERGY = '1B040033',
	PARAM_EP_RESERVE_MAX_ENERGY = '1B040034',
	EP_RESERVE = '1B800009',
	REQ_SET_EP_RESERVE = '1B000010',
}